refactor: flavor 태그  타입, 라벨 매핑 코드 리팩터링
