+++ src/shared/constants/flavorTags.ts
@@ -1,0 +1,32 @@
+export const FLAVOR_TAGS = {
+  FLOWER: ['CHRYSANTHEMUM', 'PLUM_BLOSSOM', 'ACACIA', 'LOTUS', 'ROSE'] as const,
+  FRUIT: [
+    'CITRUS',
+    'STRAWBERRY',
+    'JAPANESE_PLUM',
+    'MELON',
+    'BANANA',
+    'PEAR',
+    'PEACH',
+    'APPLE',
+    'WILD_STRAWBERRY',
+    'APRICOT',
+    'YUZU',
+    'PLUM',
+    'CANTALOUPE',
+    'PINEAPPLE',
+  ] as const,
+  GRAIN: [
+    'POTATO',
+    'FRESHLY_COOKED_RICE',
+    'SWEET_POTATO',
+    'GRAIN_POWDER',
+    'NURUNGJI',
+    'WHEAT',
+    'RAW_RICE',
+    'CORN',
+  ] as const,
+  NUT: ['PEANUT', 'CHESTNUT', 'ALMOND', 'PINE_NUT'] as const,
+  SWEETENER: ['HONEY', 'MALT_SYRUP', 'BROWN_RICE_SYRUP', 'CARAMEL'] as const,
+  DAIRY: ['BUTTER', 'YOGURT', 'MILK', 'CHEESE'] as const,
+};
